+++ src/app/app.component.ts
@@ -1,11 +1,12 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
+import { FormsModule } from '@angular/forms';
 
 import { HeaderComponent } from "./header/header.component";
 import { PokemonListComponent } from "./pokemon-list/pokemon-list.component";
 import { AuthComponent } from './auth/auth.component';
-import { AuthService } from './service/auth.service';
+import { AuthService } from './services/auth.service';
 
 
 @Component({
@@ -13,7 +14,7 @@ import { AuthService } from './service/auth.service';
     standalone: true,
     templateUrl: './app.component.html',
     styleUrl: './app.component.scss',
-    imports: [CommonModule, HeaderComponent, PokemonListComponent, AuthComponent, RouterOutlet],
+    imports: [CommonModule, HeaderComponent, PokemonListComponent, AuthComponent, RouterOutlet, FormsModule],
     providers: [AuthService]
 })
 export class AppComponent {
