+++ src/app/auth/auth.component.html
@@ -1,11 +1,12 @@
 <div class="row">
     <form>
-        <div class="form-group">
-            <label for="email">Email</label>
-            <input type="email" id="'email" class="form-control">
-        </div>
-        <div>
-            <button class="btn btn-primary">Log in</button>
-        </div>
+      <div class="form-group">
+        <label for="email">Email</label>
+        <input type="email" id="email" class="form-control" [(ngModel)]="email" [ngModelOptions]="{standalone: true}">
+      </div>
+      <div>
+        <button class="btn btn-primary" (click)="handleLogin()">Log in</button>
+      </div>
     </form>
-</div>>
+  </div>
+
