+++ src/app/auth/auth.component.scss
@@ -2,7 +2,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    height: 100vh;
+    height: 50vh;
   
     form {
       width: 300px;
