+++ src/app/auth/auth.component.ts
@@ -1,19 +1,21 @@
 // auth.component.ts
 import { Component } from '@angular/core';
-import { AuthService } from '../service/auth.service';
+import { AuthService } from '../services/auth.service';
+import { FormsModule } from '@angular/forms';
 
 @Component({
   selector: 'app-auth',
   standalone: true,
-  imports: [],
+  imports: [FormsModule],
   templateUrl: './auth.component.html',
   styleUrl: './auth.component.scss'
 })
 export class AuthComponent {
+  email: string = '';
+  
   constructor(private authService: AuthService) {}
 
-  // Use authService.loggedIn in your component
-  get loggedIn(): boolean {
-    return this.authService.loggedIn;
-  }
+handleLogin(): void {
+  this.authService.login(this.email);
+}
 }
