+++ src/app/header/header.component.html
@@ -1,5 +1,6 @@
 <!-- header.component.html -->
 <div class="header">
     <h1>Pokemon App</h1>
+    <button *ngIf="loggedIn" (click)="logout()">Logout</button>
   </div>
