+++ src/app/header/header.component.scss
@@ -11,4 +11,14 @@
     letter-spacing: 3px;
     font-weight: 700;
   }
+
+  button {
+    padding: 10px;
+    background-color: rgb(127, 28, 35);
+    color: #fff;
+    border: none;
+    border-radius: 3px;
+    cursor: pointer;
+    font-size: 16px;
+  }
 }
