+++ src/app/header/header.component.ts
@@ -1,12 +1,23 @@
 import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AuthService } from '../services/auth.service';
 
 @Component({
   selector: 'app-header',
   standalone: true,
-  imports: [],
+  imports: [CommonModule],
   templateUrl: './header.component.html',
   styleUrl: './header.component.scss'
 })
 export class HeaderComponent {
+  constructor(private authService: AuthService) {}
+
+  get loggedIn(): boolean {
+    return this.authService.loggedIn;
+  }
+
+  logout(): void {
+    this.authService.logout();
+  }
 
 }
