--- src/app/service/auth.service.ts
@@ -1,10 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AuthService {
-
-  loggedIn: boolean = true;
-  constructor() { }
-}
